requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
tensorflow==2.14.0
numpy==1.24.0
pillow==11.1.0